requirements.txt
anthropic==0.25.8
chromadb==0.4.24
fastapi==0.111.0
func-timeout==4.3.5
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-auth==2.29.0
google-generativeai==0.3.2
googleapis-common-protos==1.63.0
langchain==0.1.8
langchain-anthropic==0.1.4
langchain-chroma==0.1.0
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==0.0.9
langchain-openai==0.0.6
langgraph==0.0.25
langsmith==0.1.57
numpy==1.26.4
openai==1.28.1
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-multipart==0.0.9
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
SQLAlchemy==2.0.30
sqlglot==23.3.0
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
transformers==4.40.2
typing_extensions==4.11.0
datasketch==1.6.4


# @RK
psycopg2==2.9.9